requirements.txt
asttokens==2.2.1
backcall==0.2.0
cfgv==3.3.1
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.6
executing==1.2.0
faiss-cpu==1.7.4
filelock==3.12.0
fonttools==4.39.3
identify==2.5.24
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.23.0
ipython==8.12.2
jedi==0.18.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
nodeenv==1.7.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
pre-commit==3.3.1
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
PyYAML==6.0
pyzmq==25.0.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
stack-data==0.6.2
threadpoolctl==3.1.0
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
virtualenv==20.23.0
wcwidth==0.2.6
zipp==3.15.0
